moved tests from /spec to /__tests__
