+++ .eslintrc.js
@@ -17,11 +17,7 @@ module.exports = {
 
   overrides: [
     {
-      files: [
-        '**/__tests__/**/*.js',
-        '**/__mocks__/**/*.js',
-        'packages/acceptance-tests/**/*.js',
-      ],
+      files: ['**/__tests__/**/*.js', '**/__mocks__/**/*.js'],
       env: {
         jest: true,
         node: true,
