+++ packages/acceptance-tests/__tests__/init-oauth2.spec.js
@@ -1,6 +1,6 @@
 const http = require('http');
-const cmd = require('../helpers/cmd');
-const { CONFIG_FILE_PATH } = require('../../lib/constants');
+const cmd = require('./helpers/cmd');
+const { CONFIG_FILE_PATH } = require('../lib/constants');
 const rimraf = require('rimraf');
 const { existsSync, readFileSync } = require('fs');
 const yaml = require('js-yaml');
