+++ packages/acceptance-tests/jasmine.json
@@ -1,6 +1,6 @@
 {
-  "spec_dir": "packages/acceptance-tests/spec",
-  "spec_files": ["tests/*[sS]pec.js"],
+  "spec_dir": "packages/acceptance-tests/__tests__",
+  "spec_files": ["*[sS]pec.js"],
   "helpers": ["helpers/**/*.js"],
   "stopSpecOnExpectationFailure": false,
   "random": false
