+++ packages/acceptance-tests/run-tests
@@ -6,7 +6,7 @@ const { CONFIG_FILE_PATH } = require('./lib/constants');
 const path = require('path');
 const Jasmine = require('jasmine');
 const rimraf = require('rimraf');
-const cmd = require('./spec/helpers/cmd');
+const cmd = require('./__tests__/helpers/cmd');
 const Reporter = require('jasmine-terminal-reporter');
 const { config } = require('yargs');
 const reporter = new Reporter({
